requirements.txt
streamlit==1.32.0
openai==1.12.0
python-dotenv==1.0.0
mcp==1.6.0
langchain-mcp-adapters
langgraph
langchain-google-genai
anthropic
